Removing cache now that we're more performant
